added tabs for diffrent hubs
