+++ src/components/home/dev-pack-tabs.js
@@ -4,20 +4,36 @@ import React from 'react';
 import { TabPanel, Tabs, TabList, Tab } from 'react-tabs';
 import './dev-pack-tabs.css';
 import DevCardHub from './../hub/dev-card/dev-card-hub';
+import DomainsHub from '../hub/domain/domains-hub';
+import { DevCardStateContext } from './../../context/devcard-context';
+import AuthHeader from './../hub/auth-header';
 
 const DevPackTabs = () => {
+  const state = React.useContext(DevCardStateContext);
   return (
     <section
       sx={{
         backgroundColor: 'accent',
         minHeight: 700,
-        width: '100vw',
-        marginTop: '10em',
+        // width: '100%',
+        my: 5,
+        mb: -10,
       }}
     >
+      <header
+        sx={{
+          margin: '0 auto',
+          width: '90%',
+          mb: 6,
+        }}
+      >
+        <AuthHeader userName={state.name} />
+      </header>
+
       <div
         sx={{
-          m: 6,
+          maxWidth: 1440,
+          margin: '0 auto',
         }}
       >
         <Tabs className="react-tabs">
@@ -27,6 +43,7 @@ const DevPackTabs = () => {
                 sx={{
                   fontFamily: 'heading',
                   fontSize: 2,
+                  color: 'text',
                 }}
               >
                 Dev Card
@@ -39,7 +56,7 @@ const DevPackTabs = () => {
                   fontSize: 2,
                 }}
               >
-                Domain MGMT
+                Domains
               </p>
             </Tab>
           </TabList>
@@ -47,7 +64,9 @@ const DevPackTabs = () => {
           <TabPanel className="react-tabs__tab-panel">
             <DevCardHub />
           </TabPanel>
-          <TabPanel className="react-tabs__tab-panel">domains mgmt</TabPanel>
+          <TabPanel className="react-tabs__tab-panel">
+            <DomainsHub />
+          </TabPanel>
         </Tabs>
       </div>
     </section>
