+++ src/components/hub/domain/domains-hub.js
@@ -2,8 +2,27 @@
 import { jsx } from 'theme-ui';
 import { TabPanel } from 'react-tabs';
 
-const DomainsHub = ({ user, ...rest }) => {
-  return <TabPanel {...rest}>This is the domains panel</TabPanel>;
+const DomainsHub = () => {
+  return (
+    <div
+      sx={{
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+        minHeight: 700,
+      }}
+    >
+      <h1
+        sx={{
+          fontFamily: 'heading',
+          fontSize: 6,
+          color: 'background',
+        }}
+      >
+        COMING SOON
+      </h1>
+    </div>
+  );
 };
 
 DomainsHub.tabsRole = 'TabPanel';
